--- freecad/CADQuery/Qt/__init__.py
@@ -1,2 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileNotice: Part of the CADQuery addon.
