+++ freecad/CadQuery/Qt/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileNotice: Part of the CadQuery addon.
