+++ freecad/CadQuery/display.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: Apache-2.0
-# SPDX-FileNotice: Part of the CADQuery addon.
+# SPDX-FileNotice: Part of the CadQuery addon.
 
 # Define the show_object function which CadQuery execution environments need to provide
 def show_object(cq_object, options=None):
